test(parse): Unify indexing parse tests and add dotted index tests
